--- src/layouts/Chat.vue
@@ -1,384 +1,0 @@
-<template>
-  <q-layout
-    ref="chatLayout"
-    view="lhh LpR lff"
-    container
-    class="hide-scrollbar absolute full-width"
-  >
-    <!-- Send file dialog -->
-    <!-- TODO: Move this up.  We don't need a copy of this dialog for each address (likely) -->
-    <q-dialog v-model="sendFileOpen">
-      <send-file-dialog :address="address" />
-    </q-dialog>
-
-    <!-- Send money dialog -->
-    <q-dialog v-model="sendMoneyOpen">
-      <send-lotus-dialog
-        :address="address"
-        :contact="contactProfile"
-      />
-    </q-dialog>
-
-    <q-header>
-      <q-toolbar class="q-pl-sm">
-        <q-btn
-          class="q-px-sm"
-          flat
-          dense
-          @click="toggleMyDrawerOpen"
-          icon="menu"
-        />
-        <q-avatar rounded>
-          <img :src="contactProfile.avatar">
-        </q-avatar>
-        <q-toolbar-title class="h6">
-          {{ contactProfile.name }}
-        </q-toolbar-title>
-        <q-space />
-        <q-btn
-          class="q-px-sm"
-          flat
-          dense
-          @click="toggleContactDrawerOpen"
-          icon="person"
-        />
-      </q-toolbar>
-    </q-header>
-
-    <q-page-container>
-      <q-page>
-        <q-scroll-area
-          ref="chatScroll"
-          @scroll="scrollHandler"
-          class="q-px-none absolute full-width full-height"
-        >
-          <template v-for="({ chunk, globalIndex }, index) in chunkedMessages">
-            <chat-message-stack
-              v-if="chunk[0]"
-              :key="index"
-              :address="address"
-              :messages="chunk"
-              :global-index="globalIndex"
-              :contact="getContact(chunk[0].outbound)"
-              :name-color="chunk[0].outbound ? '': `color: ${nameColor};`"
-              @replyClicked="({ address, payloadDigest }) => setReply(payloadDigest)"
-            />
-          </template>
-        </q-scroll-area>
-        <q-page-sticky
-          position="bottom-right"
-          :offset="[18, 18]"
-          v-show="!bottom"
-        >
-          <q-icon
-            round
-            size="md"
-            name="arrow_circle_down"
-            @mousedown.prevent="scrollBottom"
-            color="accent"
-          />
-        </q-page-sticky>
-      </q-page>
-    </q-page-container>
-
-    <q-footer bordered>
-      <div
-        v-if="!!replyDigest"
-        class="reply col q-px-md q-pt-sm"
-        ref="replyBox"
-      >
-        <!-- Reply box -->
-        <div class="row">
-          <div class="col text-weight-bold">
-            {{ replyName }}
-          </div>
-          <div class="col-auto">
-            <q-btn
-              dense
-              flat
-              color="accent"
-              icon="close"
-              @click="setReply(null)"
-            />
-          </div>
-        </div>
-        <div class="row q-px-sm q-pt-sm">
-          <chat-message-text
-            v-if="replyItem.type=='text'"
-            :text="replyItem.text"
-          />
-          <chat-message-image
-            v-else-if="replyItem.type=='image'"
-            :image="replyItem.image"
-          />
-          <chat-message-stealth
-            v-else-if="replyItem.type=='stealth'"
-            :amount="replyItem.amount"
-          />
-        </div>
-      </div>
-
-      <!-- Message box -->
-      <chat-input
-        ref="chatInput"
-        v-model:message="message"
-        v-model:stamp-amount="stampAmount"
-        @sendMessage="sendMessage"
-        @sendFileClicked="sendFileOpen = true"
-        @giveLotusClicked="sendMoneyOpen = true"
-      />
-    </q-footer>
-  </q-layout>
-</template>
-
-<script>
-import { mapGetters, mapActions } from 'vuex'
-
-import ChatInput from '../components/chat/ChatInput.vue'
-import ChatMessageStack from '../components/chat/messages/ChatMessageStack.vue'
-import SendFileDialog from '../components/dialogs/SendFileDialog.vue'
-import ChatMessageText from '../components/chat/messages/ChatMessageText.vue'
-import ChatMessageImage from '../components/chat/messages/ChatMessageImage.vue'
-import ChatMessageStealth from '../components/chat/messages/ChatMessageStealth.vue'
-import SendLotusDialog from '../components/dialogs/SendLotusDialog.vue'
-import { insufficientStampNotify } from '../utils/notifications'
-import { addressColorFromStr } from '../utils/formatting'
-import { stampLowerLimit } from '../utils/constants'
-
-const scrollDuration = 0
-
-export default {
-  props: {
-    address: {
-      type: String,
-      default: () => ''
-    },
-    messages: {
-      type: Array,
-      default: () => []
-    }
-  },
-  emits: ['toggleMyDrawerOpen', 'toggleContactDrawerOpen'],
-  components: {
-    ChatMessageStack,
-    ChatMessageText,
-    ChatMessageImage,
-    ChatMessageStealth,
-    SendFileDialog,
-    ChatInput,
-    SendLotusDialog
-  },
-  data () {
-    return {
-      // TODO: Timer isn't used delete
-      timer: null,
-      sendFileOpen: false,
-      sendMoneyOpen: false,
-      bottom: true,
-      messagesToShow: 30,
-      message: '',
-      replyDigest: null
-    }
-  },
-  unmounted () {
-    clearTimeout(this.timer)
-  },
-  mounted () {
-    this.scrollBottom()
-  },
-  methods: {
-    ...mapGetters({
-      getAcceptancePrice: 'contacts/getAcceptancePrice',
-      getStampAmount: 'chats/getStampAmount'
-    }),
-    ...mapActions({
-      setStampAmount: 'chats/setStampAmount'
-    }),
-    sendMessage (message) {
-      const stampAmount = this.getStampAmount()(this.address)
-      const acceptancePrice = this.getAcceptancePrice()(this.address)
-      if (stampAmount < acceptancePrice) {
-        insufficientStampNotify()
-      }
-      if (message !== '') {
-        this.$relayClient.sendMessage({
-          address: this.address,
-          text: message,
-          replyDigest: this.replyDigest,
-          stampAmount
-        })
-        this.message = ''
-        this.replyDigest = null
-      }
-    },
-    scrollBottom () {
-      const scrollArea = this.$refs.chatScroll
-      if (!scrollArea) {
-        // Not mounted yet
-        return
-      }
-      const scrollTarget = scrollArea.getScrollTarget()
-      if (!this.bottom && scrollTarget.scrollTop >= 10) {
-        return
-      }
-      this.$nextTick(() => {
-        scrollArea.setScrollPosition(
-          'vertical',
-          scrollTarget.scrollHeight,
-          scrollDuration
-        )
-      })
-    },
-    getContact (outbound) {
-      if (outbound) {
-        return this.getProfile
-      } else {
-        return this.getContactVuex(this.address).profile
-      }
-    },
-    scrollHandler (details) {
-      if (
-        // Ten pixels from top
-        details.verticalPosition <= 10
-      ) {
-        this.messagesToShow += 30
-        return
-      }
-      // Set this afterwards, incase we were at the bottom already.
-      // We want to ensure that we scroll!
-      this.bottom = details.verticalSize - details.verticalPosition - details.verticalContainerSize <= 10
-    },
-    setReply (payloadDigest) {
-      this.replyDigest = payloadDigest
-    },
-    shouldShowHeader (message, previousMessage) {
-      if (previousMessage === undefined) {
-        return true
-      }
-      return previousMessage.senderAddress !== message.senderAddress
-    },
-    toggleMyDrawerOpen () {
-      this.$emit('toggleMyDrawerOpen')
-    },
-    toggleContactDrawerOpen () {
-      this.$emit('toggleContactDrawerOpen')
-    }
-  },
-  computed: {
-    chunkedMessages () {
-      // TODO: Improve stacking logic e.g. long durations between messages prevent stacking
-      // TODO: Optimize this by progressively constructing it
-      if (!this.messages) {
-        return []
-      }
-
-      const length = Math.min(this.messages.length, this.messagesToShow)
-      const start = this.messages.length - length
-      const end = this.messages.length
-
-      const { chunks, currentChunk, globalIndex } = this.messages.slice(start + 1, end).reduce(({ chunks, currentChunk, globalIndex }, message) => {
-        if (currentChunk[0].senderAddress === message.senderAddress) {
-          currentChunk.push(message)
-          return { chunks, currentChunk, globalIndex: globalIndex }
-        } else {
-          chunks.push({ chunk: currentChunk, globalIndex: globalIndex })
-          return { chunks, currentChunk: [message], globalIndex: globalIndex + currentChunk.length }
-        }
-      }, { chunks: [], currentChunk: [this.messages[start]], globalIndex: start })
-
-      chunks.push({ chunk: currentChunk, globalIndex: globalIndex })
-      return chunks
-    },
-    nameColor () {
-      return addressColorFromStr(this.address)
-    },
-    ...mapGetters({
-      getContactVuex: 'contacts/getContact',
-      getMessageByPayload: 'chats/getMessageByPayload',
-      getProfile: 'myProfile/getProfile'
-    }),
-    replyItem () {
-      if (!this.replyDigest) {
-        return null
-      }
-      const msg = this.getMessageByPayload(this.replyDigest)
-      if (!msg) {
-        return null
-      }
-
-      const firstNonReply = msg.items.find(item => item.type !== 'reply')
-
-      // Focus input box
-      this.$refs.chatInput.focus()
-
-      return firstNonReply
-    },
-    replyName () {
-      const replyAddress = this.getMessageByPayload(this.replyDigest).senderAddress
-      if (replyAddress === this.$wallet.myAddress.toXAddress()) {
-        return this.getProfile.name
-      }
-      const contact = this.getContactVuex(replyAddress)
-      return contact.profile.name
-    },
-    replyColor () {
-      const replyAddress = this.getMessageByPayload(this.replyDigest).senderAddress
-      if (replyAddress === this.$wallet.myAddress.toXAddress()) {
-        return 'black'
-      }
-
-      return this.addressColorFromStr(this.address)
-    },
-    contactProfile () {
-      return this.getContactVuex(this.address).profile
-    },
-    stampAmount: {
-      set (stampAmount) {
-        const stampAmountNumber = Math.max(stampLowerLimit, Number(stampAmount))
-        this.setStampAmount({ address: this.address, stampAmount: stampAmountNumber })
-      },
-      get () {
-        return Number(this.getStampAmount()(this.address))
-      }
-    }
-  },
-  watch: {
-    'messages.length' () {
-      // Scroll to bottom if user was already there.
-      this.scrollBottom()
-    },
-    active (newActive) {
-      if (!newActive) {
-        return
-      }
-
-      // Focus input box
-      this.$refs.chatInput.focus()
-      // Scroll to bottom only if the view was effectively in it's initial state.
-      this.scrollBottom()
-      // TODO: Scroll to last unread
-    }
-  }
-}
-</script>
-
-<style lang="scss" scoped>
-.reply {
-  padding: 5px 0;
-  color: var(--q-color-text);
-  background: var(--q-color-background);
-  padding-left: 8px;
-  border-left: 3px;
-  border-left-style: solid;
-  border-left-color: $primary;
-}
-
-.scroll-area-bordered {
-  border-right: 1px;
-  border-right-style: solid;
-  border-right-color: $separator-color;
-  border-bottom: 1px;
-  border-bottom-style: solid;
-  border-bottom-color: $separator-color;
-}
-</style>
